generalize console watching to all common errors
