+++ test/utils/react.jsx
@@ -1,7 +1,5 @@
 import React from 'react';
 import TestUtils from 'react-addons-test-utils';
-import sinon from 'sinon';
-import _ from 'underscore';
 import { ParentComponentWithContext } from './parent-component';
 
 export function scryRenderedDOMComponentsWithId(tree, id) {
@@ -56,29 +54,4 @@ export function wrapComponentWithContext(Component, props, context) {
     );
 
     return wrapper.getWrappedInstance();
-}
-
-export function watchConsole() {
-    sinon.spy(console, 'error'); // ensure that we don't swallow up console error output
-
-    sinon.spy(console, 'warn'); // ensure that we don't swallow up console warn output
-}
-
-export function getReactWarningErrorArray() {
-    const errors = _.times(console.error.callCount, (index) => console.error.getCall(index));
-    const warnings = _.times(console.warn.callCount, (index) => console.warn.getCall(index));
-
-    return errors.concat(warnings)
-        .filter((c) => (c.args && c.args.length && /(Invalid|Failed)/gi.test(c.args[0])))
-        .map((c) => c.args.join('|'));
-}
-
-export function unwatchConsole() {
-    if (console.error.restore) {
-        console.error.restore();
-    }
-
-    if (console.warn.restore) {
-        console.warn.restore();
-    }
 }
